requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.2
filelock==3.0.12
idna==3.3
mysql-connector-python==8.0.28
platformdirs==2.3.0
pymongo==4.0.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
virtualenv==20.7.2
